requirements_streamlit.txt
streamlit
pyyaml
torch==1.8.0
torchvision==0.9.0
scipy
captum
scikit-image
scikit-learn
termcolor
tqdm
opencv-python